pyproject.toml
[tool.poetry]
name = "speech-synth"
version = "0.1.0"
description = ""
authors = ["camille-004 <dunningcamille@gmail.com>"]
readme = "README.md"
packages = [{include = "speech_synth"}]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
pydantic = "^2.7.1"
django-allauth = "^0.63.2"
drf-yasg = "^1.21.7"
PyJWT = "^2.8.0"
cryptography = "^42.0.7"
mysqlclient = "^2.2.4"
six = "^1.16.0"
types-six = "^1.16.21.20240513"
pillow = "^10.3.0"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
django-stubs = "^5.0.0"
djangorestframework-stubs = "^3.15.0"
djhtml = "^3.0.6"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
mypy_path = ["stubs"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"